+++ example/lib/main.dart
@@ -1,9 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:ui_extension/ui_extension.dart';
+// import 'package:ui_extension/ui_extension.dart';
 
 void main() {
   runApp(const MainApp());
 }
+
 const s1 = Duration(seconds: 1);
 
 class MainApp extends StatelessWidget {
@@ -11,6 +13,7 @@ class MainApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    Material;
     return MaterialApp(
       home: Scaffold(
         body: Center(
@@ -20,27 +23,74 @@ class MainApp extends StatelessWidget {
               // NestClipRRect(),
               // NestPhysicalModel(),
             ],
-            child: Column(
-              children: [
-                const Text('data')
-                    .ui()
-                    // .rounded(borderRadius: BorderRadius.circular(55))
-                    // .physical(elevation: 5)
-                    .padding(all: 50)
-                    .bordered(
-                        border: GradientBorder.all(
-                      gradient: const LinearGradient(
-                        colors: [
+            child: SingleChildScrollView(
+              child: Column(
+                children: [
+                  Material(
+                    elevation: 50,
+                    child: Container(
+                      height: 300,
+                      width: 200,
+                      color: Colors.white,
+                    ),
+                  ),
+                  const Gap(15),
+                  Container(
+                    height: 300,
+                    width: 200,
+                    color: Colors.white,
+                  ).ui().bordered(
+                        radius: 95,
+                        depth: -1,
+                        elevation: 5,
+                        width: 5,
+                        color: Colors.red,
+                        gradient: const SweepGradient(colors: [
                           Colors.red,
+                          Colors.orange,
+                          Colors.yellow,
+                          Colors.green,
                           Colors.blue,
-                        ],
+                          Colors.indigo,
+                          Colors.purple,
+                          Colors.red,
+                        ]),
+                        // width: 15,
+                        duration: const Duration(seconds: 1),
                       ),
-                    ))
-                // .paddingValue(all: 10)
-                // .clipRRect(borderRadius: BorderRadius.circular(56))
-                // .decorated(const BoxDecoration(color: Colors.red))
-                // .sized(dimension: 150, duration: 300.ms)
-              ],
+                  const Gap(15),
+                  UiBordered(
+                    border: Border.all(
+                      width: 5,
+                      strokeAlign: -5,
+                      // gradient: const LinearGradient(colors: [
+                      //   Colors.red,
+                      //   Colors.orange,
+                      //   Colors.yellow,
+                      //   Colors.green,
+                      //   Colors.blue,
+                      //   Colors.indigo,
+                      //   Colors.purple,
+                      // ]),
+                    ),
+                    borderRadius: BorderRadius.circular(65),
+                    child: Container(
+                      // color: Colors.red,
+                      height: 200,
+                      width: 300,
+                      color: Colors.red,
+                    ),
+                  ),
+                  Container(
+                    color: Colors.red,
+                    height: 300,
+                    width: 300,
+                  )
+                      .ui() //
+                      .constrained(maxHeight: 150)
+                      .bordered(radius: 50, depth: -1, duration: s1)
+                ],
+              ),
             ),
           ),
         ),