+++ lib/ui_extension.dart
@@ -1,18 +1,17 @@
 library ui_extension;
 
 
+// Ui utils
+export 'package:gap/gap.dart';
+export 'package:bordered/bordered.dart';
+export 'package:animated_value/animated_value.dart';
+// Ui extensions
+export 'package:ui_extension/extensions/context_extension.dart';
+export 'package:ui_extension/extensions/element_extension.dart';
+export 'package:ui_extension/extensions/ui_extensions.dart';
 // Ui core
 export 'package:ui_extension/src/ui_list.dart';
 export 'package:ui_extension/src/ui_manager.dart';
 export 'package:ui_extension/widgets/ui.dart';
 export 'package:ui_extension/widgets/ui_widgets.dart';
-export 'package:ui_extension/extensions/ui_extensions.dart';
-
-// Ui extensions
-export 'package:ui_extension/extensions/context_extension.dart';
-export 'package:ui_extension/extensions/element_extension.dart';
-
-// Ui utils
-export 'package:gap/gap.dart';
-export 'package:ui_extension/utils/gradient_border.dart';