+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Changelog
+
+## 0.1.0
+
+* Added UiBordered through `bordered` library (exported).
+* Removed UiRoundedBox, UiClipRRect, and UiPhysicalModel. Use UiBordered instead.
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial release.